% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExprLinePlot.R
\name{ExprLinePlot}
\alias{ExprLinePlot}
\title{Gene expression levels lineplot}
\usage{
ExprLinePlot(dataset = df, genelist = c("GAPDH", "CAPN6"))
}
\arguments{
\item{dataset}{Gene expression dataset with a column for gene symbols (symbol) and one column for each sample in the form "neuron_stage" can be left NA (default) and a prompt will ask for file}

\item{genelist}{Character string of gene of interest}
}
\value{
A list containing a lineplot for each gene
}
\description{
This functions create one lineplot for each gene provided from the provided dataset
}
\author{
Matteo Miotto
}
