% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRAnalysis.R
\name{PCRAnalysis}
\alias{PCRAnalysis}
\title{PCR Analysis Tool}
\usage{
PCRAnalysis(inputfile = NA, outputfile = NA, max_rep_diff = 0.8, housekeeping = "GADPH", sheetname = "Analysis")
}
\arguments{
\item{inputfile}{Character string of input file, can be left NA (default) and a prompt will ask for file}

\item{outputfile}{Character string of output file, if left NA (default) sheet is added to input file}

\item{max_rep_diff}{Maximum range of replicates values}

\item{housekeeping}{Character string of the housekeeping gene (GADPH as default) against which all deltaCT are calculated}

\item{sheetname}{Character string of the output sheetname (Analysis as default)}
}
\value{
An xlsx sheet with analysis performed (deltaCT and foldchange over housekeeping gene for each sample)
}
\description{
This functions takes as input the results file from a PCR analysis and gives back an analysis sheet
}
\author{
Matteo Miotto
}
