% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCRAnalysis.R
\name{PCRAnalysis}
\alias{PCRAnalysis}
\title{PCR Analysis Tool}
\usage{
PCRAnalysis(inputfile = NA, outputfile, max_rep_diff = 0.8, housekeeping = "GADPH")
}
\arguments{
\item{inputfile}{Character string of input file, can be left NA (default) and a prompt will ask for file}

\item{outputfile}{Character string of output file}

\item{max_rep_diff}{Maximum range of replicates values}

\item{housekeeping}{Character string of the housekeeping gene (GADPH as default) against which all deltaCT are calculated}
}
\value{
An csv file with analysis performed (deltaCT and foldchange over housekeeping gene for each sample)
}
\description{
This functions takes as input the results file from a PCR analysis and gives back an analysis file
}
\author{
Matteo Miotto
}
